firewall: Only allow link-local telnet connection
So far, on a vanilla LFF ipv6 telnet connections via the WAN port were
the only allowed way to access the router.

IPv6 global prefixes could have made such a router accessible via the
global internet easily, if the local internet connection provided IPv6
addresses and if the local network were connected to the router's WAN
port.

With this commit we are restricting the telnet access to IPv6 link local
only.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>
Reported-by: yanosz