tinc: Add one VPN entry node
To get connected to Lübeck's mesh network more easily, the public key of
one VPN entry node will be added in the firmware already.

To decentralize this process, a wiki page at
http://luebeck.freifunk.net/ will be maintained, where anyone can add
their own VPN entry to a list.

I will add any of the VPN entry nodes from the wiki page to the firmware
if the following criterias are met:

a) The VPN node has a reasonable latency (< 100ms)
b) The VPN node has a reasonable throughput (fast ADSL)
c) The VPN node is in Lübeck and therefore has the potential to make the
   VPN obsolete again one day when a larger wireless mesh cloud exists.
d) The owner provides a way of (pseudonymous) contact. Just to have a
   way to contact this person in case of issues.

In case of harming one of these requirements or seeing activities
harmful to the mesh integrity, I will remove such a key again.

VPNs are / will always only be seen as a way to boost the initial
deployment - however it should always be the goal of a Freifunk
community to _not_ rely on a centrally controlled, commercial medium
- that will always come back like a boomerang one day otherwise.

If you disagree with these policies - feel free to fork / alter the
repository and make your own releases. It's a git repository, so a
_decentral_ version control system (these are just my personal
policies!)

Signed-off-by: Linus Lüssing <linus.luessing@web.de>