network: Rename uci section name from 'wan' to 'private'
For some people WAN has been a little confusing, as it is not really a
"wide area network" behind this port. Also we are soon going to bridge a
couple more interfaces with it, therefore it will not been WAN only.

It is going to serve a private network, therefore renaming it to
'private'

Signed-off-by: Linus Lüssing <linus.luessing@web.de>