wireless: Adding private WPA secured Wifi interface
This commit adds another wifi interface in AP mode, bridged with the WAN
port / private network. This allows the user to also use the same device
for own, local network usage, with no need to have another, separate
device for the home network.

This WiFi interface is secured with WPA and a random 63 character
passphrase which can be looked up or changed in the LuCI web interface
(or command line, of course).

Signed-off-by: Linus Lüssing <linus.luessing@web.de>