tinc: Enable DirectOnly and disable Forwarding
To strengthen the intra city VPN's and therefore overall security, we
are enabling the DirectOnly option and disable Forwarding to encourage
people to exchange their public keys and to gain kind of a web of trust.

Nevertheless data packets can usually still be forwarded indirectly due
to the mesh routing protocol.

People that exchanged keys directly will usually have a faster
connection due to the direct connection.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>