network: Fix hostname for some targets
E.g. the MAC printed on the back of the Dlink DIR-300 is not exactly the
one of eth0. It is actually the one of ath0/wifi0, but we do not have
that information available yet during uci-defaults. Therefore we are
adding a customization of the MAC for this profile. There might be more
profiles later where the MAC is not exactly the one printed on the
router, so using the new /etc/lff_profile seems to make sense to stay
flexible.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>