lff: Adding script for safe upgrade from v0.1/v0.2rc1 to v0.2
With this commit a v0.1 or v0.2rc1 firmware can be upgraded to v0.2 if
"Keep configuration files" has been selected in the webinterface or
sysupgrade. tinc keys will be kept from v0.2rc1, and for v0.1 too if
doing a "echo '/etc/tinc/'  > /lib/upgrade/keep.d/tinc" before flashing.

Any changes to:
* /etc/config/network
* /etc/config/wireless
* /etc/config/batman-adv
* /etc/config/uhttpd
* /etc/config/tinc
* /etc/firewall.user
* /etc/opkg.conf

will nevertheless be overwritten due to compatibility reasons. Backups
of the old files can be found in /usr/backups.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>